[Python API] Add new ov Python API

+++ model-optimizer/extensions/ops/loop.py
@@ -1,26 +1,14 @@
-"""
- Copyright (C) 2017-2020 Intel Corporation
+# Copyright (C) 2018-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
 
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-"""
 import logging as log
 
 import numpy as np
 
 from extensions.ops.tensor_iterator import TensorIterator
-from mo.front.common.partial_infer.utils import int64_array
+from mo.front.common.partial_infer.utils import shape_array, is_fully_defined, dynamic_dimension_value
 from mo.graph.graph import Node, Graph
-from mo.graph.port import Port
+from mo.middle.passes.fusing.helpers import common_bfs
 from mo.middle.passes.infer import partial_infer
 from mo.ops.const import Const
 
@@ -74,6 +62,27 @@ class Loop(TensorIterator):
         return suitable_nodes[0] if len(suitable_nodes) == 1 else None
 
     @staticmethod
+    def get_external_nodes_by_internal_id(loop_node: Node, internal_layer_id: int) -> list:
+        """
+        Get a list of nodes from the main graph that are connected with a node with internal_layer_id
+        from the body graph
+
+        :param loop_node: The Loop node
+        :param internal_layer_id: Internal layer ID of the node in the body graph
+        :return: A list of external nodes (from the main graph) that are connected with a node with
+        internal_layer_id from the body graph
+        """
+        for map_item in loop_node.input_port_map:
+            if map_item['internal_layer_id'] == internal_layer_id \
+                    and loop_node.is_in_port_connected(map_item['external_port_id']):
+                return [loop_node.in_port(map_item['external_port_id']).get_source().node]
+        for map_item in loop_node.output_port_map:
+            if map_item['internal_layer_id'] == internal_layer_id \
+                    and loop_node.is_out_port_connected(map_item['external_port_id']):
+                return [dest.node for dest in loop_node.out_port(map_item['external_port_id']).get_destinations()]
+        return []
+
+    @staticmethod
     def updated_body_parameters_shape(loop_node: Node):
         """
         Update shape for Loop body parameters.
@@ -90,11 +99,14 @@ class Loop(TensorIterator):
             if body_node is not None:
                 assert body_node.soft_get('type') == 'Parameter'
 
-                input_shape = int64_array([])  # this is a current iteration number input shape
+                input_shape = shape_array([])  # this is a current iteration number input shape
                 loop_port_idx = record['external_port_id']
                 if loop_port_idx != -1:
                     input_shape = loop_node.in_port(loop_port_idx).get_connection().get_source().data.get_shape()
-                body_node.shape = input_shape
+                slice_axis = record['axis']
+                body_node.shape = input_shape.copy()
+                if slice_axis is not None:
+                    body_node.shape[slice_axis] = 1
                 log.debug('Updated shape for the body node with internal_id "{}" with value {}'
                           ''.format(record['internal_layer_id'], body_node.shape))
 
@@ -117,18 +129,13 @@ class Loop(TensorIterator):
             loop_port_idx = record['external_port_id']
             if loop_port_idx != -1:  # the id = -1 for execution condition output which is not connected anywhere
                 output_value = body_node.in_port(0).data.get_value()
-                output_shape = body_node.in_port(0).data.get_shape()
+                output_shape = body_node.in_port(0).data.get_shape().copy()
                 concat_axis = record['axis']
                 if concat_axis is not None:
                     assert output_shape[concat_axis] == 1, 'Dimension for concatenation is not equal to 1 for scan ' \
                                                            'output for Loop node "{}" for loop output port "{}"'.\
                         format(loop_name, loop_port_idx)
-                    num_iters = Loop.iterations_count(loop_node)
-                    if num_iters is None:
-                        log.error('Dynamic number of iterations for Loop node "{}". Consider number to be 1 to be able'
-                                  ' to generate the IR.'.format(loop_name), extra={'is_warning': True})
-                        num_iters = 1
-                    output_shape[concat_axis] = num_iters
+                    output_shape[concat_axis] = Loop.iterations_count(loop_node)
                 # MO does not support evaluation of Loop scan outputs with const values
                 if concat_axis is None and output_value is not None:
                     loop_node.out_port(loop_port_idx).data.set_value(output_value)
@@ -141,20 +148,25 @@ class Loop(TensorIterator):
         Try to determine the number of loop iterations. If we detect that the number is dynamic then return None.
 
         :param loop_node: Loop operation node
-        :return: number of iterations or None if the number depends on runtime values.
+        :return: number of iterations or dynamic_dimensions if the number depends on runtime values.
         """
         assert loop_node.soft_get('type') == 'Loop'
 
         if loop_node.is_in_port_connected(1):
             execution_condition = loop_node.in_port(1).data.get_value()
-            if execution_condition is None:  # dynamic execution condition
-                return None
+            if not is_fully_defined(execution_condition):  # dynamic execution condition
+                return dynamic_dimension_value
             execution_condition = execution_condition.item()
             if not execution_condition:  # 0 iterations
                 return 0
         num_iterations = loop_node.in_port(0).data.get_value()
+        if not is_fully_defined(num_iterations):
+            return dynamic_dimension_value
         if num_iterations is not None:
             num_iterations = num_iterations.item(0)
+            # in some ONNX models the num_iterations input is equal to max(int64) meaning dynamic number of iterations
+            if num_iterations < 0 or num_iterations == np.iinfo(np.int64).max:
+                return dynamic_dimension_value
         return num_iterations
 
     @staticmethod
@@ -247,42 +259,53 @@ class Loop(TensorIterator):
         return result_nodes[0]
 
     @staticmethod
-    def connect_body_input(loop_input_port: Port, internal_parameter: Node, external_node_out_port: Port = None,
+    def connect_body_input(loop_node: Node, loop_input_port_idx: int, body_parameter: Node,
                            axis: [int, None] = None, start: [int, None] = None, end: [int, None] = None,
                            stride: [int, None] = None, part_size: [int, None] = None):
-        loop_node = loop_input_port.node
-        assert loop_node.soft_get('op') == 'Loop'
-        assert loop_input_port.type == 'in'
-        assert internal_parameter.soft_get('op') == 'Parameter'
-        assert internal_parameter.id in loop_node.body
+        """
+        Update the input port map to connect the input port with the specified body parameter
 
-        if external_node_out_port is not None:
-            assert loop_input_port.disconnected()
-            assert external_node_out_port.node.id not in loop_node.body
-            loop_input_port.connect(external_node_out_port)
+        :param loop_node: the Loop node
+        :param loop_input_port_idx: the input port index to connect
+        :param body_parameter: the body parameter node to connect
+        :param axis: dimension for input slicing
+        :param start: start value of dimension from which to start slicing
+        :param end: end value of dimension when to finish slicing
+        :param stride: a step value for slicing
+        :param part_size: a partial size for slicing, i.e. slicing [start; start + part_size)
+        :return: None
+        """
+        assert loop_node.soft_get('op') == 'Loop'
+        assert body_parameter.soft_get('op') == 'Parameter'
+        assert body_parameter.id in loop_node.body
 
         loop_node.input_port_map.append({'axis': axis, 'stride': stride, 'part_size': part_size, 'start': start,
-                                         'end': end, 'external_port_id': loop_input_port.idx,
-                                         'internal_layer_id': internal_parameter['internal_layer_id']})
+                                         'end': end, 'external_port_id': loop_input_port_idx,
+                                         'internal_layer_id': body_parameter['internal_layer_id']})
 
     @staticmethod
-    def connect_body_output(loop_output_port: Port, internal_result: Node, external_node_input_ports: list = None,
-                            axis: [int, None] = None, start: [int, None] = None, end: [int, None] = None,
-                            stride: [int, None] = None, part_size: [int, None] = None):
-        loop_node = loop_output_port.node
+    def connect_body_output(loop_node: Node, loop_output_port_idx: int, internal_result: Node, axis: [int, None] = None,
+                            start: [int, None] = None, end: [int, None] = None, stride: [int, None] = None,
+                            part_size: [int, None] = None):
+        """
+        Update the output port map to connect the body Result node with the specified output port
+
+        :param loop_node: the Loop node
+        :param loop_output_port_idx: the output port index to connect
+        :param internal_result: the body Result node to connect
+        :param axis: dimension for output concatenation
+        :param start: start value of dimension from which to start concatenation
+        :param end: end value of dimension when to finish concatenation
+        :param stride: a step value for concatenation
+        :param part_size: a partial size for concatenation, i.e. concatenation [start; start + part_size)
+        :return: None
+        """
         assert loop_node.soft_get('op') == 'Loop'
-        assert loop_output_port.type == 'out'
         assert internal_result.soft_get('op') == 'Result'
         assert internal_result.id in loop_node.body
 
-        if external_node_input_ports is not None:
-            assert loop_output_port.disconnected()
-            assert all([port.node.id not in loop_node.body for port in external_node_input_ports])
-            for port in external_node_input_ports:
-                port.disconnect()
-                loop_output_port.connect(port)
         loop_node.output_port_map.append({'axis': axis, 'stride': stride, 'part_size': part_size, 'start': start,
-                                          'end': end, 'external_port_id': loop_output_port.idx,
+                                          'end': end, 'external_port_id': loop_output_port_idx,
                                           'internal_layer_id': internal_result['internal_layer_id']})
 
     @staticmethod
@@ -298,14 +321,51 @@ class Loop(TensorIterator):
                                      'to_port': 0})
 
     @staticmethod
+    def parameter_unchanged_after_iteration(loop_node: Node, body_parameter: Node):
+        """
+        Checks if the body Parameter node is connected to some body Result and the data provided to Result is not
+        changed between iterations. The data is considered unchanged if:
+        1. There is no back edge for this Parameter OR
+        2. There is a back edge from some Result to Parameter and there are only Identity ops in between or
+           Parameter is connected to Result directly.
+
+        :param loop_node: the Loop node to check
+        :param body_parameter: the body Parameter node
+        :return: the result of the check
+        """
+        assert body_parameter.id in loop_node.body
+        assert body_parameter.soft_get('op') == 'Parameter'
+        if not any([attr['to_layer'] == body_parameter.soft_get('internal_layer_id') for attr in loop_node.back_edges]):
+            return True
+
+        for back_edge_attrs in loop_node.back_edges:
+            if back_edge_attrs['to_layer'] == body_parameter.soft_get('internal_layer_id'):
+                result_internal_id = back_edge_attrs['from_layer']
+                result_nodes = loop_node.body.get_op_nodes(internal_layer_id=result_internal_id)
+                assert len(result_nodes) == 1, 'There should be exactly one node with id {}, but there are {}' \
+                                               ''.format(result_internal_id, len(result_nodes))
+                result_node = result_nodes[0]
+                # check that the Result node consumes data from Parameter node directly or through Identity operations
+                parameters = common_bfs(result_node, ['Identity'], ['Parameter'], is_backward=True, attr_to_check='op',
+                                        follow_multi_consumer_data_nodes=True)
+                if any([node.soft_get('internal_layer_id') == body_parameter.internal_layer_id for node in parameters]):
+                    return True
+        return False
+
+    @staticmethod
     def pull_constant_inputs_into_body(loop_node: Node):
         for port_idx, in_port in reversed(loop_node.in_ports().items()):
-            # TODO add a check that the input does not correspond to execution_condition
-            if not in_port.disconnected() and in_port.get_source().node.soft_get('type') == 'Const':
+            if port_idx > 1 and not in_port.disconnected() and in_port.get_source().node.soft_get('type') == 'Const':
+                body_parameter = Loop.external_port_id_to_body_node(loop_node, port_idx, loop_node.input_port_map)
+                # if there is a back edge into a body Parameter then we cannot replace it with a Const if the value
+                # is updated during each iteration. So we need to check that the tensor is passed to the next iteration
+                # unchanged
+                if not Loop.parameter_unchanged_after_iteration(loop_node, body_parameter):
+                    continue
+
                 original_const_node = in_port.get_source().node
                 new_const_node = Const(loop_node.body, original_const_node.attrs()).create_node()
 
-                body_parameter = Loop.external_port_id_to_body_node(loop_node, port_idx, loop_node.input_port_map)
                 body_parameter.out_port(0).get_connection().set_source(new_const_node.out_port(0))
                 loop_node.body.remove_nodes_from([body_parameter.id])
                 loop_node.delete_input_port(port_idx)
@@ -317,8 +377,56 @@ class Loop(TensorIterator):
             if record[attr] == original_value:
                 record[attr] = new_value
                 matched += 1
-        assert matched == 1, 'More than one record in the portmap for attr "{}" wil original value "{}"' \
+        assert matched == 1, 'More than one record in the portmap for attr "{}" with original value "{}"' \
                              ''.format(attr, original_value)
+
+    @staticmethod
+    def update_port_map_value_ext(port_map: dict, layer_id_attr: str, layer_id_value: int,
+                                  updated_attr: str, new_attr_value: int):
+        """
+        Updates a value of requested attribute for a certain layer id in a port map
+        :param port_map: a map of external ports to internal layer ids
+        :param layer_id_attr: layer id attribute for which to update attribute
+        :param layer_id_value: layer id value for which to update attribute
+        :param updated_attr: a name of attribute which to update
+        :param new_attr_value: new value of attribute
+        """
+        matched = 0
+        for record in port_map:
+            if record.get(layer_id_attr) == layer_id_value:
+                record[updated_attr] = new_attr_value
+                matched += 1
+        assert matched == 1, 'More than one record in the portmap for attr "{}" with original value "{}"' \
+                             ''.format(layer_id_attr, layer_id_value)
+
+    @staticmethod
+    def back_edge_exists(back_edges_map: dict, from_layer: int, to_layer: int):
+        """
+        Checks if a back edge exists in the back_edges_map connecting specific nodes
+        :param back_edges_map: a map where to search for specified back edge
+        :param from_layer: id of Result node that belongs a back edge
+        :param to_layer: id of Parameter node that belongs a back edge
+        :return: True or False
+        """
+        for back_edge in back_edges_map:
+            if back_edge['from_layer'] == from_layer and back_edge['to_layer'] == to_layer:
+                return True
+        return False
+
+    @staticmethod
+    def inter_edge_exists(port_map: dict, external_port_id: int, internal_layer_id: int):
+        """
+        Check if inter-graph edge (i.e. an edge between the main graph and body graph) exists
+        :param port_map: a port map where to search for inter-graph edge
+        :param external_port_id: port index from/to which edge goes
+        :param internal_layer_id: layer id from/to which edge goes
+        :return: True or False
+        """
+        for i_port in port_map:
+            if i_port['external_port_id'] == external_port_id and \
+                    i_port['internal_layer_id'] == internal_layer_id:
+                return True
+        return False
 
     @staticmethod
     def re_numerate_input_ports(loop_node: Node):
@@ -344,7 +452,8 @@ class Loop(TensorIterator):
                     new_port_id += 1
 
             for port_idx_to_remove in reversed(range(new_port_id, max_port_id + 1)):
-                loop_node.delete_input_port(port_idx_to_remove)
+                if port_idx_to_remove in loop_node.in_ports().keys():
+                    loop_node.delete_input_port(port_idx_to_remove)
 
     @staticmethod
     def re_numerate_output_ports(loop_node: Node):
@@ -372,7 +481,8 @@ class Loop(TensorIterator):
                     new_port_id += 1
 
             for port_idx_to_remove in reversed(range(new_port_id, max_port_id + 1)):
-                loop_node.delete_output_port(port_idx_to_remove)
+                if port_idx_to_remove in loop_node.out_ports().keys():
+                    loop_node.delete_output_port(port_idx_to_remove)
 
     @staticmethod
     def remove_unused_ops_from_port_map(loop_node: Node, port_map: dict, port_map_attr: str, dir: [None, str] = None):
@@ -399,7 +509,8 @@ class Loop(TensorIterator):
                 port_to_remove = port_map[record_id_to_remove]['external_port_id']
                 if port_to_remove != -1:
                     if dir == 'in':
-                        if port_to_remove not in [0, 1]:  # input port 0 and 1 are mandatory for the Loop node
+                        # input port 0 and 1 are mandatory for the Loop node
+                        if port_to_remove not in [0, 1] and port_to_remove in loop_node.in_ports().keys():
                             loop_node.delete_input_port(port_to_remove)
                     elif dir == 'out' and port_to_remove in loop_node.out_ports():
                         loop_node.delete_output_port(port_to_remove)
